SquashFS
Новый формат игровых архивов. Лучше сжатие, быстрее распаковка/упаковка, нет лимита по размеру архива. Архивы нового формата должны иметь расширение `.sq*`.
Для упаковки: https://github.com/pmq20/squashfuse/files/691217/sqfs43-win32.zip
```cmd
sqfs43-win32\mksquashfs.exe gamedata gamedata.sq_01 -comp lz4 -Xhc
pause
```
Для распаковки подойдёт https://github.com/mcmilk/7-Zip-zstd/releases/download/v24.09-v1.5.7-R1/7z24.09-zstd-x64.exe
-------------------------------------
Так же добавлена поддержка Anomaly-style архивов в подкаталогах. Дополнительные каталоги для поиска можно прописать в `fsgame.ltx`:
```ini
$app_arch_root$         = true|   false|  $fs_root$		  | game_archs\
```

+++ ogsr_engine/xrCore/FS.cpp
@@ -134,7 +134,7 @@ u32 IWriter::chunk_size() // returns size of currently opened chunk, 0 otherwise
 void IWriter::w_compressed(void* ptr, u32 count, const bool encrypt, const bool is_ww)
 {
     BYTE* dest = 0;
-    unsigned dest_sz = 0;
+    size_t dest_sz = 0;
     _compressLZ(&dest, &dest_sz, ptr, count);
 
     if (encrypt)
@@ -192,7 +192,7 @@ IReader* IReader::open_chunk(u32 ID)
         if (bCompressed)
         {
             BYTE* dest;
-            unsigned dest_sz;
+            size_t dest_sz;
             _decompressLZ(&dest, &dest_sz, pointer(), dwSize);
             return xr_new<CTempReader>(dest, dest_sz, tell() + dwSize);
         }
@@ -243,7 +243,7 @@ IReader* IReader::open_chunk_iterator(u32& ID, IReader* _prev)
     {
         // compressed
         u8* dest;
-        unsigned dest_sz;
+        size_t dest_sz;
         _decompressLZ(&dest, &dest_sz, pointer(), _size);
         return xr_new<CTempReader>(dest, dest_sz, tell() + _size);
     }
